requirements.txt
playwright>=1.40.0
gspread>=5.12.0
oauth2client>=4.1.3
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
aiohttp>=3.9.0